Cargo: update dependencies, fix some dependency related security issues.
== h2 vulnerable to denial of service ==

Hyper is an HTTP library for Rust and h2 is an HTTP 2.0 client & server
implementation for Rust. An issue was discovered in h2 v0.2.4 when
processing header frames. It incorrectly processes the HTTP2 RST_STREAM
frames by not always releasing the memory immediately upon receiving the
reset frame, leading to stream stacking. As a result, the memory and CPU
usage are high which can lead to a Denial of Service (DoS).

This issue affects users only when dealing with http2 connections.

== spin-rs initialisation failure in `Once::try_call_once` can lead to
   undefined behaviour for other initialisers ==

Once::try_call_once is unsound if invoked more than once concurrently
and any call fails to initialise successfully.

Signed-off-by: Silvio Fricke <silvio.fricke@gmail.com>